Applying metrics plugin to graphql.
