+++ opencti-platform/opencti-graphql/package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.34.0",
+    "@bmatei/apollo-prometheus-exporter": "^2.1.1",
     "@elastic/elasticsearch": "7.13.0",
     "@graphql-tools/graphql-file-loader": "^7.3.1",
     "@graphql-tools/load": "^7.4.0",
@@ -95,6 +96,7 @@
     "passport-ldapauth": "3.0.1",
     "passport-local": "1.0.0",
     "passport-saml": "3.1.1",
+    "prom-client": "^14.0.1",
     "python-shell": "3.0.0",
     "ramda": "0.27.1",
     "redlock": "4.2.0",
