+++ opencti-platform/opencti-graphql/src/http/httpServer.js
@@ -5,7 +5,7 @@ import http from 'http';
 // noinspection NodeCoreCodingAssistance
 import { readFileSync } from 'fs';
 import conf, { booleanConf, logApp } from '../config/conf';
-import createApp from './httpPlatform';
+import createApp, { applyWildcard } from './httpPlatform';
 import createApolloServer from '../graphql/graphql';
 import { isStrategyActivated, STRATEGY_CERT } from '../config/providers';
 
@@ -16,8 +16,9 @@ const CERT_KEY_CERT = conf.get('app:https_cert:crt');
 const CA_CERTS = conf.get('app:https_cert:ca');
 const rejectUnauthorized = booleanConf('app:https_cert:reject_unauthorized', true);
 const createHttpServer = async () => {
-  const apolloServer = createApolloServer();
-  const { app, seeMiddleware } = await createApp(apolloServer);
+  const { app, seeMiddleware } = await createApp();
+  const apolloServer = createApolloServer(app);
+  applyWildcard(app); // Needed in order to register prometheus metrics
   let httpServer;
   if (CERT_KEY_PATH && CERT_KEY_CERT) {
     const key = readFileSync(CERT_KEY_PATH);
